fix(explorer): move outputs into main route group
