+++ apps/explorer/app/(main)/output/[id]/loading.tsx
@@ -1,0 +1,5 @@
+import { OutputSkeleton } from '../../../../components/OutputSkeleton'
+
+export default function Loading() {
+  return <OutputSkeleton />
+}
