+++ apps/explorer/app/(main)/output/[id]/not-found.tsx
@@ -1,4 +1,4 @@
-import { StateError } from '../../../components/StateError'
+import { StateError } from '../../../../components/StateError'
 
 export default function Page() {
   return (
