+++ apps/explorer/app/(main)/output/[id]/opengraph-image.tsx
@@ -2,9 +2,9 @@ import { truncate } from '@siafoundation/design-system'
 import { to } from '@siafoundation/request'
 import { humanSiacoin } from '@siafoundation/units'
 
-import { getOGImage } from '../../../components/OGImageEntity'
-import { getExplored } from '../../../lib/explored'
-import { ExplorerPageProps } from '../../../lib/pageProps'
+import { getOGImage } from '../../../../components/OGImageEntity'
+import { getExplored } from '../../../../lib/explored'
+import { ExplorerPageProps } from '../../../../lib/pageProps'
 
 export const revalidate = 0
