--- apps/explorer/app/output/[id]/loading.tsx
@@ -1,5 +1,0 @@
-import { OutputSkeleton } from '../../../components/OutputSkeleton'
-
-export default function Loading() {
-  return <OutputSkeleton />
-}
